refactor: BookSearchBottomSheet 컴포넌트 코드 정리 및 구조화
